--limit now works, except for CV. (testing still needed)

+++ include/timbl/MBLClass.h
@@ -135,8 +135,8 @@ namespace Timbl {
     size_t MaxFeats() const { return MaxFeatures; };
     int Progress() const { return progress; };
     void Progress( int p ){ progress =  p; };
-    std::string extract_limited_m( int );
-    Target   *Targets;
+    std::string extract_limited_m( size_t );
+    Target *Targets;
     std::vector<Feature *> Features;
     std::vector<Feature *> PermFeatures;
     std::vector<size_t> permutation;
